==> R/01_load.qmd <==
---
title: "01_load"
format: html
editor: visual
---

In this file, we're loading the raw data, and we write it as .tsv in the /data folder.

## Loading libraries

```{r loadLibraries_01}
#| message: false
library("tidyverse")
```

## Loading data

```{r loadData_01}
first_path <- "~/projects/R_project"
file_path <- "data/_raw/Data_Cortex_Nuclear.csv" 
file_path <- file.path(first_path, 
                       file_path)

df <- read.csv(file_path)
```

```{r writeData_01}
file_path <- "data/01_dat_load.tsv"
file_path <- file.path(first_path, 
                       file_path)

df |>
  write_tsv(file = file_path)
```

The eight classes of mice are described based on features such as genotype, behavior, and treatment. According to genotype, mice can be control or trisomic. According to behavior, some mice have been stimulated to learn (context-shock) and others have not (shock-context) and in order to assess the effect of the drug memantine in recovering the ability to learn in trisomic mice, some mice have been injected with the drug and others have not.
