==> R/04_describe.qmd <==
---
title: "04_describe"
format: html
editor: visual
---

In this file, we create a linear model between 2 variables, and we try to analyze if there are differences in the modelling, between the 8 groups (Genotype = 0 or 1, Treatment = 0 or 1, Stimulated = 0 or 1).

## Loading libraries

```{r loadLib_04}
#| message: false
library("tidyverse")
library("broom")
library("purrr")
```

## Loading data

```{r loadData_04}
first_path <- "~/projects/R_project"
file_path <- "data/03_dat_aug.tsv" 
file_path <- file.path(first_path, 
                       file_path)
df_aug <- read_tsv(file = file_path)
```

## **Creating the model**

We first group by three categories: Genotype (control = 0 or Ts65Dn = 1), Treatment (Memantine = 1 or Saline = 0), and Stimulated (yes = 1, no = 0)

```{r model1}
df_nested <- df_aug |>
  group_by(Genotype, 
           Treatment, 
           Stimulated) |>
  nest()
```

Then we create a linear model between the two variables BRAF_N and pERK_N.

```{r model2}
df_models <- df_nested |>
  mutate(model = map(data, 
                     ~lm(BRAF_N ~ pERK_N, 
                         data = .x))  )
```

Here we get the coefficients.

```{r model3}
df_results <- df_models |>
  mutate(model_summary = map(model, 
                             tidy)) |>
  unnest(model_summary)
```

Showing the results.

```{r model4}
df_results |>
  select(Genotype, 
         Treatment, 
         Stimulated, 
         term, 
         estimate, 
         std.error, 
         p.value)
```

Then, we decided to add the q.value and the variable is_significant. We set the threshold to 0.01 because it is the highest (reasonable) value that creates a distinction between significant and non-significant values. With a higher threshold (e.g., 5%), there are no differences between significant and non-significant values, as all values are classified as significant.

```{r model5}
df_estimates <- df_results |>
  mutate(q.value = p.adjust(p.value),
         is_significant = case_when(q.value <= 0.01 ~ "yes",
                                    q.value > 0.01 ~ "no"))
```

```{r model6}
df_final <- df_estimates |>
  unnest(data)
```

### Final plot to analyze differences between the groups

We first need a new variable we will be using in the plot showing which group we're studying.

```{r model7}
df_final <- df_final |>
  mutate(Group = paste0("Genotype: ", Genotype, 
                        ", Treatment: ", Treatment, 
                        ", Stimulated: ", Stimulated))
```

Then we can do the plot, showing the 8 groups, each with the two bars coming from the statistics: Intercept (bar on the left of each group subplot), and pERK_N (bar on the right of each group subplot).

```{r, model8, fig.width=14, fig.height=12}
df_final |>
  ggplot(mapping = aes(x = term,
                       y = estimate, 
                       fill = is_significant)) +
  geom_bar(stat = "identity", 
           position = "dodge") +
  facet_wrap(~ Group) +
  labs(title = "Coefficient estimates for each group",
       x = "Predictive variable: Intercept (Left), pERK_N (Right)",
       y = " ") +
  scale_x_discrete(labels = c("(Intercept)" = "", 
                              "pERK_N" = "")) +
  theme_minimal() +
  theme(
    plot.title = element_text(size = 25, 
                              face = "bold"), 
    axis.title = element_text(size = 16),                
    axis.text = element_text(size = 14),                
    legend.title = element_text(size = 14),             
    legend.text = element_text(size = 12),             
    strip.text = element_text(size = 14)                 
  ) 
```

The plot shows the coefficient estimates for the two predictive variables, **Intercept** and **pERK_N**, across different groups defined by combinations of Genotype, Treatment and Stimulated. The x-axis represents the predictive variables, and the bars represent their estimated coefficients, colored by whether they are statistically significant.

### Key observations:

1.  **Intercept for (0,0,1)**:

    -   The only bar marked as **not significant** is the Intercept bar for the (0,0,1) group (Control mice, treated with saline, and stimulated). This coefficient has the smallest value among all groups, approaching zero, which indicates a very weak or negligible effect for this combination of Genotype, Treatment and Stimulated.

2.  **Significance and Comparison of pERK_N**:

    -   All the **pERK_N** bars, are marked as significant, indicating that the effect of **pERK_N** is statistically noteworthy across all combinations of the experimental groups.

    -   The **pERK_N** coefficients show varying magnitudes depending on the group. In some groups, like (0,0,0) and (0,0,1), the effect is stronger, suggesting a more pronounced influence of **pERK_N**, while in other groups, like (1,0,0) the effect is weaker but still significant.

3.  **Comparing Intercept and pERK_N**:

    -   In all the groups, the **pERK_N** coefficient is larger than the **Intercept**, suggesting that **pERK_N** has a stronger influence on the outcome compared to the baseline represented by the **Intercept**.

    -   The magnitude of the coefficient for **pERK_N** is higher in **Genotype 0 (Control)** groups compared to **Genotype 1 (Trisomic)**, indicating that the effect of **pERK_N** may be modulated by the genotype, with control mice showing a more substantial effect.

### 

In conclusion, the plot provides insights into the influence of **pERK_N** and **Intercept** across various experimental conditions. The **Intercept** is only significant in specific cases, while **pERK_N** shows consistent significance across groups, with its effect varying in magnitude depending on the combination of genotype and treatment. The **Genotype 0** groups generally show larger effects, particularly in the case of no stimulation.
