==> R/05_analysis_1.qmd <==
---
title: "05_analysis_1"
format: html
editor: visual
---

Before going into descriptive statistics, we analyze the general trend of the data to understand what we're dealing with.

## Loading libraries

```{r loadLib_05}
#| message: false
library("tidyverse")
```

## Loading data

```{r loadData_05}
first_path <- "~/projects/R_project"
file_path <- "data/03_dat_aug.tsv" 
file_path <- file.path(first_path, 
                       file_path)

df_aug <- read_tsv(file = file_path)
```

## Analysis

First, we count the occurance of each class in the dataset. The classes are the following:

1.  **c-CS-s**: control mice, stimulated to learn, injected with saline

2.  **c-CS-m**: control mice, stimulated to learn, injected with memantine

3.  **c-SC-s**: control mice, not stimulated to learn, injected with saline

4.  **c-SC-m**: control mice, not stimulated to learn, injected with memantine

5.  **t-CS-s**: trisomy mice, stimulated to learn, injected with saline

6.  **t-CS-m**: trisomy mice, stimulated to learn, injected with memantine

7.  **t-SC-s**: trisomy mice, not stimulated to learn, injected with saline

8.  **t-SC-m**: trisomy mice, not stimulated to learn, injected with memantine

```{r}
ggplot(data = df_aug,
       mapping = aes(x = class, 
                     fill = class)) +
  geom_bar(alpha = 0.5) +
  scale_fill_manual(values = rainbow(8)) +
  theme_light(base_family = "Avenir") +
  labs(title = "Count of each class")
```

The classes related to control mice appear to have a slightly higher count compared to the classes related to trisomic mice, suggesting an imbalance in the data between the two groups. The imbalanced distribution could be significant, but not necessarily problematic, as it reflects a true disparity between the groups (more mice in the control group). So, for this study we won't apply techniques to address balancing.

Going into details, we pick the two proteins that we used to create the model in 04_describe (BRAF_N and pERK_N), and we analise their expression levels based on the class.

```{r}
ggplot(data = df_aug,
       mapping = aes(x = class,
                     y = BRAF_N,
                     fill = class)) +
  geom_boxplot(alpha = 0.7) +
  scale_fill_manual(values = rainbow(8)) +
  theme_light(base_family = "Avenir") +
  scale_y_continuous(limits = c(0, 3.75)) +
  labs(title = "Distribution of BRAF_N by class") +
  theme(legend.position = "none")

ggplot(data = df_aug,
       mapping = aes(x = class,
                     y = pERK_N,
                     fill = class)) +
  geom_boxplot(alpha = 0.7) +
  scale_fill_manual(values = rainbow(8)) +
  theme_light(base_family = "Avenir") +
  scale_y_continuous(limits = c(0, 3.75)) +
  labs(title = "Distribution of pERK_N by class") +
  theme(legend.position = "none")
```

We may notice that just a few of the boxplots are similar between the two proteins.

The expression levels of BRAF_N appear to be more stable and less variable, especially in the control group. Instead, pERK_N shows greater variability across classes, particularly in the trisomic group: this is represented by larger boxes, longer whiskers and a higher number of outliers. However, they both present most of the outliers in c-CS-s and a very small variance in c-SC-m and c-SC-s.
