==> R/07_analysis_3.qmd <==
---
title: "07_analysis_3"
format: html
editor: visual
---

In this file we analyse specific protein level distributions separating the samples by genotype.

## Installing packages

```{r installPatchwork}
#install.packages("patchwork")
```

## Loading libraries

```{r loadLib_07}
#| message: false
library("tidyverse")
library("patchwork")
```

## Loading data

```{r loadData_07}
first_path <- "~/projects/R_project"
file_path <- "data/03_dat_aug.tsv" 
file_path <- file.path(first_path, 
                       file_path)
df_aug <- read_tsv(file = file_path)
```

## Analysis

We're going to visualize the distribution of BRAF_N and pERK_N for the control and trisomic groups separately.

```{r analysis_31}
trisomic <- df_aug |>
  filter(Genotype == 1)

control <- df_aug |>
  filter(Genotype == 0)
```

```{r analysis_32}
#| eval: false
# Unsuccessful attempt to use a function to plot the densitograms:
source(file = "10_densitogram.R")
```

```{r analysis_33}
#| eval: false
# Unsuccessful attempt to use a function to plot the densitograms:
p1 <- densitogram(trisomic, 
                  "BRAF_N", 
                  deparse(substitute(trisomic)))
p1
```

```{r analysis_34}
p1 <- ggplot(data = trisomic,
             mapping = aes(x = BRAF_N,
                           fill = class)) +
  geom_density(alpha = 0.5) +
  labs(title = "Densitogram of BRAF_N for trisomic classes") +
  theme_minimal(base_family = "Avenir")

p2 <- ggplot(data = control,
             mapping = aes(x = BRAF_N,
                           fill = class)) +
  geom_density(alpha = 0.5) + 
  labs(title = "Densitogram of BRAF_N for control classes") +
  theme_minimal(base_family = "Avenir")

p1 / p2
```

We set the same axis in order to better compare the two groups.

```{r analysis_35}
p1 <- p1 +
  xlim(0.1, 0.9) +
  ylim(0, 11.0)

p2 <- p2 +
  xlim(0.1, 0.9) +
  ylim(0, 11.0)

p1 / p2
```

For both groups the stimulation classes (CS, red and green) show main peaks around intermediate BRAF_N values (approximately 0.4), and display flatter distributions compared to the non-stimulation classes (SC, blue and purple) classes, indicating greater variability in the data. On the other hand, the non-stimulation classes are mainly overlapped and concentrated around lower values, between 0.1 and 0.3, with a relatively sharp density.

This clear separation between the presence and the absence of stimulation suggests that the protein expression levels may be distinctive for these categories.

The non-stimulation classes are dominant in the control group but not in the trisomic group, suggesting a possible biological effect. In general, trisomic classes appear to have more diverse distributions than controls, which was also noticed in the first analysis.

```{r analysis_36}
p3 <- ggplot(data = trisomic,
             mapping = aes(x = pERK_N,
                           fill = class)) +
  geom_density(alpha = 0.5) +
  labs(title = "Densitogram of pERK_N for trisomic classes") +
  theme_minimal(base_family = "Avenir")

p4 <- ggplot(data = control,
             mapping = aes(x = pERK_N,
                           fill = class)) +
  geom_density(alpha = 0.5) + 
  labs(title = "Densitogram of pERK_N for control classes") +
  theme_minimal(base_family = "Avenir")

p3 / p4
```

We set the same axis in order to better compare the two groups.

```{r analysis_37}
p3 <- p3 +   
  xlim(0.1, 1.5) +   
  ylim(0, 10.0)  

p4 <- p4 +   
  xlim(0.1, 1.5) +   
  ylim(0, 10.0)  

p3 / p4
```

For both groups the stimulation classes (CS, red and green) show main peaks around intermediate pERK_N values (approximately 0.6), and display flatter distributions compared to the non-stimulation classes (SC, blue and purple) classes, indicating greater variability in the data. On the other hand, the non-stimulation classes are concentrated around lower values, between 0.2 and 0.5, and show higher peaks.

This clear separation between the presence and the absence of stimulation suggests that the protein expression levels may be distinctive for these categories, just as BRAF_N.

Unlike the previous one, in this case the SC-m class (no stimulation, treated with memantine) is dominant in the trisomic group as it is in the control group. The main difference is represented by the SC-s class (no stimulation, treated with saline), which is dominant in the control group but not in the trisomic group, suggesting a possible biological effect.

In general, trisomic classes appear to have more diverse distributions than controls, which was also noticed for BRAF_N and in the first analysis as well. The differences in the distribution of pERK_N between the two groups could suggest alterations in the phosphorylation levels of the ERK protein, potentially linked to differences in the molecular mechanisms between the two conditions.

In conclusion, this analysis reveals distinct protein expression patterns between stimulation and non-stimulation classes for both BRAF_N and pERK_N. Stimulation classes show greater variability, while non-stimulation classes are more concentrated at lower levels.

Key differences between control and trisomic groups also suggest potential biological effects, with trisomic groups consistently displaying more diverse distributions.
